requirements.txt
matplotlib==3.4.3
numpy==1.21.2
scipy==1.7.1
streamlit==1.2.0
Pillow==8.4.0
opencv-python==4.10.0.84
moviepy==1.0.3
mutagen==1.47.0
